+++ data/gaphics/gfx.py
@@ -1,0 +1,48 @@
+def gameTitle():
+    print("                _______________________________________________________________________________")
+    print("              / \                                                                              \.")
+    print("              |  |                                                                             |.")
+    print("               \_|               ████████▄   ███    █▄     ▄████████    ▄████████     ███      |.")
+    print("                 |              ███    ███  ███    ███   ███    ███   ███    ███ ▀█████████▄   |.")
+    print("                 |              ███    ███  ███    ███   ███    █▀    ███    █▀     ▀███▀▀██   |.")
+    print("                 |              ███    ███  ███    ███  ▄███▄▄▄       ███            ███   ▀   |.")
+    print("                 |     ___ _ _  ███    ███  ███    ███ ▀▀███▀▀▀     ▀███████████     ███       |.")
+    print("                 |    | . | | | ███    ███  ███    ███   ███    █▄           ███     ███       |.")
+    print("                 |    |  _|_  | ███  ▀ ███  ███    ███   ███    ███    ▄█    ███     ███       |.")
+    print("                 |    |_| |___| ▀██████▀▄█ ████████▀    ██████████  ▄████████▀     ▄████▀      |.")
+    print("                 |   __________________________________________________________________________|___")
+    print("                 |  /                                                                             /.")
+    print("                 \_/_____________________________________________________________________________/.")
+
+
+def splash():
+    print("                           .,,uod8B8bou,,.")
+    print("              ..,uod8BBBBBBBBBBBBBBBBRPFT?l!i:.")
+    print("         ,=m8BBBBBBBBBBBBBBBRPFT?!||||||||||||||")
+    print("         !...:!TVBBBRPFT||||||||||!!^^""'     ||||")
+    print("         !.......:!?|||||!!^^""'              ||||")
+    print("         !.........||||                     ||||")
+    print("         !.........||||      Jackson's      ||||")
+    print("         !.........||||                     ||||")
+    print("         !.........||||          CLI        ||||")
+    print("         !.........||||                     ||||")
+    print("         !.........||||         Games       ||||")
+    print("         `.........||||,                    ||||")
+    print("          .;.......||||               _.-!!|||||")
+    print("   .,uodWBBBBb.....||||        _.-!!|||||||||!:'")
+    print("!YBBBBBBBBBBBBBBb..!|||:..-!!|||||||!iof68BBBBBb....")
+    print("!..YBBBBBBBBBBBBBBb!!||||||||!iof68BBBBBBRPFT?!::   `.")
+    print("!....YBBBBBBBBBBBBBBbaaitf68BBBBBBRPFT?!:::::::::     `.")
+    print("!......YBBBBBBBBBBBBBBBBBBBRPFT?!::::::;:!^:`;:::       .")
+    print("!........YBBBBBBBBBBRPFT?!::::::::::^''...::::::;         iBBbo.")
+    print("`..........YBRPFT?!::::::::::::::::::::::::;iof68bo.      WBBBBbo.")
+    print("  `..........:::::::::::::::::::::::;iof688888888888b.     `YBBBP^'")
+    print("    `........::::::::::::::::;iof688888888888888888888b.     `")
+    print("      `......:::::::::;iof688888888888888888888888888888b.")
+    print("        `....:::;iof688888888888888888888888888888888899fT!")
+    print("          `..::!8888888888888888888888888888888899fT|!^*'")
+    print("            `' !!988888888888888888888888899fT|!^*'")
+    print("                `!!8888888888888888899fT|!^*'")
+    print("                  `!988888888899fT|!^*'")
+    print("                    `!9899fT|!^*'")
+    print("                      `!^*'")
